Fixed FileNotFoundError
Fixed an error when History export file does not already exist

+++ history.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import os
+from os.path import exists
 
 class History:
     def __init__(self, filename="history"):
@@ -13,23 +14,24 @@ class History:
         self.history = pd.concat([self.history, data])
         
     def export_history(self):
-        os.remove(self.filename + ".xlsx")
+        file_exists = exists(self.filename + ".xlsx")
+        if file_exists: os.remove(self.filename + ".xlsx")
         self.history.to_excel(self.filename + ".xlsx", sheet_name='a', index = False)
 
 # for testing
 
-# obj = History()
-# if(not obj.get_history().empty): print(str(obj.get_history()))
+obj = History("history2")
+if(not obj.get_history().empty): print(str(obj.get_history()))
 
-# data = {
-#     "Time": 0.5,
-#     "Exchange": "FTX",
-#     "Path": [['a','b']],
-#     "Profitability": 0.1
-# }
-# df = pd.DataFrame(data)
-# obj.append_history(df)
-# obj.append_history(df)
-# print(str(obj.get_history()))
-# obj.export_history()
-# print("History exported")
+data = {
+    "Time": 0.5,
+    "Exchange": "FTX",
+    "Path": [['a','b']],
+    "Profitability": 0.1
+}
+df = pd.DataFrame(data)
+obj.append_history(df)
+obj.append_history(df)
+print(str(obj.get_history()))
+obj.export_history()
+print("History exported")
